03_API_SAMPLE/api_base/requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose
passlib[bcrypt]
sqlalchemy
requests
pytest
httpx